Modernize (#100)
* Upgrade documentation and example to openjdk8

Java 7 support is now deprecated and openjdk8 is now the recommended
environment.  Update the documentation accordingly and also bring the
Java example into compliance.

* Upgrade to the 1.9.73 AppEngine SDK

This is primarily to support the openjdk8 environment (in a subsequent
commit), as Java 7 has been deprecated.

* Upgrade to the 3.1.0 javax.servet API

Use a more modern javax.servlet API so that the official examples for the
openjdk8 environment can work.  In particular, this is to support the
WebServlet annotation, which appears in the documentation, and thus the
removal of the web.xml sample file.